rename fast_thread_local -> thread_local_dtor; thread_local -> thread_local_key
