+++ src/libstd/sys/windows/thread_local_key.rs
