+++ src/libstd/sys_common/mod.rs
@@ -65,7 +65,8 @@ pub mod remutex;
 pub mod rwlock;
 pub mod thread;
 pub mod thread_info;
-pub mod thread_local;
+pub mod thread_local_key;
+pub mod thread_local_dtor;
 pub mod util;
 pub mod wtf8;
